customized for gcc 5.4

+++ nodes/ipt/gateway/src/controller.cpp
@@ -287,11 +287,11 @@ namespace node
 						cyng::param_factory("stopbits", "one"),	//	one, onepointfive, two
 						cyng::param_factory("speed", 115200),
 
-						cyng::param_factory(sml::OBIS_W_MBUS_PROTOCOL.to_str(), mbus::S_MODE),	//	0 = T-Mode, 1 = S-Mode, 2 = S/T Automatic
+						cyng::param_factory(sml::OBIS_W_MBUS_PROTOCOL.to_str(), static_cast<std::uint8_t>(mbus::S_MODE)),	//	0 = T-Mode, 1 = S-Mode, 2 = S/T Automatic
 						cyng::param_factory(sml::OBIS_W_MBUS_S_MODE.to_str(), 30),	//	seconds
 						cyng::param_factory(sml::OBIS_W_MBUS_T_MODE.to_str(), 20),	//	seconds
 						cyng::param_factory(sml::OBIS_W_MBUS_REBOOT.to_str(), 0),	//	0 = no reboot
-						cyng::param_factory(sml::OBIS_W_MBUS_POWER.to_str(), mbus::STRONG),	//	low, basic, average, strong (unused)
+						cyng::param_factory(sml::OBIS_W_MBUS_POWER.to_str(), static_cast<std::uint8_t>(mbus::STRONG)),	//	low, basic, average, strong (unused)
 						cyng::param_factory(sml::OBIS_W_MBUS_INSTALL_MODE.to_str(), true),	//	install mode
 
 						cyng::param_factory("transparent-mode", false),
@@ -896,7 +896,8 @@ namespace node
 			//
 			//	get wireless M-Bus default configuration
 			//
-			auto const radio_protocol = cyng::numeric_cast<std::uint8_t>(dom["wired-LMN"].get(sml::OBIS_W_MBUS_PROTOCOL.to_str()), mbus::S_MODE);
+// 			auto const radio_protocol = cyng::numeric_cast<std::uint8_t>(dom["wired-LMN"].get(sml::OBIS_W_MBUS_PROTOCOL.to_str()), static_cast<std::uint8_t>(1));
+			auto const radio_protocol = cyng::numeric_cast<std::uint8_t>(dom["wired-LMN"].get(sml::OBIS_W_MBUS_PROTOCOL.to_str()), static_cast<std::uint8_t>(mbus::S_MODE));
 			CYNG_LOG_INFO(logger, sml::OBIS_W_MBUS_PROTOCOL.to_str() << " (OBIS_W_MBUS_PROTOCOL): " << +radio_protocol);
 			config.insert("_Config", cyng::table::key_generator(sml::OBIS_W_MBUS_PROTOCOL.to_str()), cyng::table::data_generator(radio_protocol), 1, tag);
 
@@ -912,7 +913,8 @@ namespace node
 			CYNG_LOG_INFO(logger, sml::OBIS_W_MBUS_REBOOT.to_str() << " (OBIS_W_MBUS_REBOOT): " << reboot << " seconds, " << (reboot / 3600) << " h");
 			config.insert("_Config", cyng::table::key_generator(sml::OBIS_W_MBUS_REBOOT.to_str()), cyng::table::data_generator(reboot), 1, tag);
 
-			auto const radio_power = cyng::numeric_cast<std::uint8_t>(dom["wired-LMN"].get(sml::OBIS_W_MBUS_POWER.to_str()), mbus::STRONG);
+// 			auto const radio_power = cyng::numeric_cast<std::uint8_t>(dom["wired-LMN"].get(sml::OBIS_W_MBUS_POWER.to_str()), static_cast<std::uint8_t>(3));
+			auto const radio_power = cyng::numeric_cast<std::uint8_t>(dom["wired-LMN"].get(sml::OBIS_W_MBUS_POWER.to_str()), static_cast<std::uint8_t>(mbus::STRONG));
 			CYNG_LOG_INFO(logger, sml::OBIS_W_MBUS_POWER.to_str() << " (OBIS_W_MBUS_POWER): " << +radio_power);
 			config.insert("_Config", cyng::table::key_generator(sml::OBIS_W_MBUS_POWER.to_str()), cyng::table::data_generator(radio_power), 1, tag);
 
@@ -922,6 +924,5 @@ namespace node
 
 
 		}
-		
 	}
 }
